tcp: remove redundant check on tskb
[ Upstream commit d1edc085559744fbda7a55e97eeae8bd6135a11b ]

The non-null check on tskb is always false because it is in an else
path of a check on tskb and hence tskb is null in this code block.
This is check is therefore redundant and can be removed as well
as the label coalesc.

if (tsbk) {
        ...
} else {
        ...
        if (unlikely(!skb)) {
                if (tskb)       /* can never be true, redundant code */
                        goto coalesc;
                return;
        }
}

Addresses-Coverity: ("Logically dead code")
Signed-off-by: Colin Ian King <colin.king@canonical.com>
Reviewed-by: Mukesh Ojha <mojha@codeaurora.org>
Signed-off-by: Eric Dumazet <edumazet@google.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Stable-dep-of: 94062790aedb ("tcp: defer shutdown(SEND_SHUTDOWN) for TCP_SYN_RECV sockets")
Signed-off-by: Sasha Levin <sashal@kernel.org>
(cherry picked from commit 6f7082e284ec3fa32694e5704da510194229ef12)
[Vegard: fix conflicts in context due to missing commits
 e2080072ed2d98a55ae69d95dea60ff7a17cddd5 ("tcp: new list for sent but
 unacked skbs for RACK recovery") and
 75c119afe14f74b4dd967d75ed9f57ab6c0ef045 ("tcp: implement rb-tree based
 retransmit queue").]
Signed-off-by: Vegard Nossum <vegard.nossum@oracle.com>
Signed-off-by: Abdallah Ibrahim <iabdollah40@gmail.com>